cosas

+++ ProyectoMagicolor/Datos/DIngreso.cs
@@ -42,8 +42,8 @@ namespace Datos
             set { _Factura = value; }
         }
 
-        private double _Impuesto;
-        public double impuesto
+        private int _Impuesto;
+        public int impuesto
         {
             get { return _Impuesto; }
             set { _Impuesto = value; }
@@ -159,7 +159,7 @@ namespace Datos
 
         }
 
-        public DIngreso(int IdIngreso, int IdTrabajador, int IdProveedor, DateTime Fecha, string Factura, double Impuesto, int MetodoPago, int Estado)
+        public DIngreso(int IdIngreso, int IdTrabajador, int IdProveedor, DateTime Fecha, string Factura, int Impuesto, int MetodoPago, int Estado)
         {
             this.idIngreso = IdIngreso;
             this.idTrabajador = IdTrabajador;
